Addition of login page, signup page.
I added a function called loadSite() in script-main.js . For this function to work properly, it should be called as the event handler for when the body loads. In other words, the body should be written with <body onload="loadSite(x)">. More information is written in the documentation, but the number that is passed to the loadSite function should indicate how many folders to go up to reach the folder that has the index.html file. So the homepage (index.html) should call this function with level set to 0 (because index.html is in the same folder as itself).

The links don't really work yet, so that last paragraph isn't really that important for now.