Phase 1 of combining order pages
Links are now set to allow the user to navigate through the website.

For now all the functionality is on the Dairy Free order page, but soon it will be added to all pages.

A quote was copied-and-pasted from the powerpoint to the homepage so it's not just a blank section.